Make align prop of Badges compatible with other components GH-374
