+++ packages/components/src/badges.js
@@ -2,18 +2,17 @@ import React from 'react';
 import { arrayOf, oneOf, string } from 'prop-types';
 import { Badge, Box, Flex } from 'theme-ui';
 
+import { responsiveAlignType } from './types';
+import { normalizeAlign } from './util';
+
 // Add margin around each badge.
 // Neutralize outer margins with negative margins on Flex.
-const Badges = ({
-  values,
-  color,
-  bg,
-  align = 'flex-start',
-  variant,
-  ...props
-}) => (
+const Badges = ({ values, color, bg, align = 'start', variant, ...props }) => (
   <Box {...props}>
-    <Flex sx={{ flexWrap: 'wrap', justifyContent: align }} m={-1}>
+    <Flex
+      sx={{ flexWrap: 'wrap', justifyContent: normalizeAlign(align) }}
+      m={-1}
+    >
       {values.map(value => (
         <Badge key={value} sx={{ color, bg, m: 1 }} variant={variant}>
           {value}
@@ -27,7 +26,7 @@ Badges.propTypes = {
   values: arrayOf(string).isRequired,
   color: string.isRequired,
   bg: string.isRequired,
-  align: oneOf(['flex-start', 'center', 'flex-end']),
+  align: responsiveAlignType,
   variant: oneOf(['outline']),
 };
